requirements.txt
faster-whisper
edge-tts
numpy
scipy
PyAudio
aiohttp
requests
PyYAML
